Head Gate indicator로 구현 l0_regularizer와 다르게 0,1 binary

+++ src/model/concrete_gate.py
@@ -52,7 +52,7 @@ class ConcreteGate(nn.Module):
 
         gates = self.get_gates()
 
-        gates = gates > 0.5 # pruning : https://www.aclweb.org/anthology/P19-1580/
+        gates = (gates > 0.5).float() # pruning : https://www.aclweb.org/anthology/P19-1580/
 
         penalty = self.get_penality()
